Image standard sizing
Implmented a function to provide standard sizing for the Image 
Component.  Options are tiny, small, medium (default), and large.  
Function will return an object with height and width for the appropriate 
size.  Currently, the size dimensions are hard coding in this function. 
Future extension may obtain these from application prarameters.