Adding the option to not resize Image (accept warping)
If prop fix=standard, the standard width and height will be accepted. 
This may warp the image.  

I used this for small avatars next to people's name.  I want to 
guarantee for the layout the avatars are the same size.  With such a 
small image, the warp effect is not huge.