Bug fix: Create form initialization for user fields
Subtle bug - If there are 2 fields that reference nwn_user (user in the 
future), only the second was defaulted with the user context.  This was 
because immutability was helper and used the original state while 
looping around the fields.

Fixed as this is preparation for adding the ability to pass in any form 
value as a prop