bugfix: CreateForm, when user has to log i and there was no field that referenced nwn_user, page crashed
The state variable that prevented the initialized method from running 
again was not triggered unless there was a nwn_user referenced field.  
In the future, this should be cleaned up to not take 2 passes through 
intialize proc (we know we need the context and do not have it). 
Anticipate doing this when we add context variables other that nwn_user