Feature - if a dynamic component is available, will load that into the body
Over time, the rest of the Body components will be added to the 
functional_components object and be dyanically loaded

This was a proof of concept that will continue to evolve.   

Need to figure out how to automatically pass all the appropriate props 
and refactor to include all the possible components.  Dynamic loading 
will work with class and functional components so the naming will 
evolve.  Trying it first with the newer functional components)

Great news, as I was really wanting to figure this out for a long time.  
This will allow new components to be added in client specific librarys 
without changing the core code.